requirements.txt
absl-py==2.1.0
accelerate==0.34.2
adagio==0.2.6
aiodns==3.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
arrow==1.3.0
ast-comments==1.2.2
async-timeout==5.0.1
attributedict==0.3.0
attrs==24.2.0
autogluon.common==1.2
autogluon.core==1.2
autogluon.features==1.2
autogluon.multimodal==1.2
autogluon.tabular==1.2
autogluon.timeseries==1.2
beautifulsoup4==4.12.3
binaryornot==0.4.4
bitarray==3.0.0
blessings==1.7
blis==1.0.1
boto3==1.36.1
botocore==1.36.1
Bottleneck==1.4.2
Brotli==1.0.9
cached-property==2.0.1
cachetools==5.5.0
catalogue==2.0.10
catboost==1.2.7
ccxt==4.4.46
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
clickhouse-connect==0.8.9
clickhouse-driver==0.2.9
cloudpathlib==0.20.0
cloudpickle==3.1.1
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.6
colorlog==6.9.0
colour-runner==0.1.1
confection==0.1.4
contourpy==1.3.1
cookiecutter==2.6.0
coreforecast==0.0.12
coverage==7.6.9
cryptography==44.0.0
cycler==0.12.1
cymem==2.0.11
cysystemd==1.6.2
cytoolz==1.0.0
dataclasses==0.8
datasets==2.2.1
deepdiff==8.0.1
defusedxml==0.7.1
dill==0.3.9
distlib==0.3.9
dnspython==2.7.0
eth_abi==5.1.0
eth-account==0.13.4
eth-bloom==3.0.1
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-tester==0.12.0b2
eth-typing==5.0.1
eth-utils==5.1.0
evaluate==0.4.1
fastai==2.7.18
fastapi==0.115.6
fastcore==1.7.20
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.16.1
fonttools==4.55.3
freqtrade==2024.12
freqtrade-client==2024.12
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.12.0
fugue==0.9.1
future==1.0.0
gdown==5.2.0
gluonts==0.16.0
gmpy2==2.1.5
google-api-core==2.24.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.62.2
h11==0.14.0
h2==4.1.0
hexbytes==1.2.1
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
huggingface_hub==0.27.1
humanfriendly==10.0
humanize==4.11.0
hyperframe==6.0.1
hyperopt==0.2.7
idna==3.10
imagecodecs==2024.9.22
imageio==2.36.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
inspecta==0.1.3
janus==2.0.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langcodes==3.4.1
language_data==1.3.0
lazy_loader==0.4
lightgbm==4.4.0
lightning==2.5.0.post0
lightning-utilities==0.11.9
linkify-it-py==2.0.3
llvmlite==0.43.0
lru-dict==1.3.0
lz4==4.3.3
Mako==1.3.8
marisa-trie==1.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.12.0
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mlforecast==0.13.4
model-index==0.1.11
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.17
munkres==1.1.4
murmurhash==1.0.10
networkx==2.8.8
nlpaug==1.1.11
nltk==3.9.1
numba==0.60.0
numexpr==2.10.1
numpy==1.26.4
omegaconf==2.3.0
opencensus==0.11.3
opencensus-context==0.1.3
openmim==0.3.7
optuna==4.1.0
ordered-set==4.1.0
orderly-set==5.2.2
orjson==3.10.13
packaging==24.2
pandas==2.2.3
pandas_ta==0.3.14b0
parsimonious==0.10.0
patsy==1.0.1
pdf2image==1.17.0
pillow==11.0.0
pip==24.2
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
pprintpp==0.4.0
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.3
psutil==6.1.1
psycopg2-binary==2.9.10
py-ecc==7.0.1
py-evm==0.10.1b2
py_find_1st==1.1.7
py4j==0.10.9.9
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycares==4.5.0
pycoingecko==3.2.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.19.1
PyJWT==2.10.1
Pympler==1.1
pyOpenSSL==24.3.0
pyparsing==3.2.1
pyproject-api==1.8.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-rapidjson==1.20
python-slugify==8.0.4
python-telegram-bot==21.10
python-whois==0.9.5
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.3.0
pytz==2024.1
pyu2f==0.1.5
pyunormalize==16.0.0
PyWavelets==1.8.0
PyYAML==6.0.2
questionary==2.1.0
ray==2.37.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
responses==0.18.0
rich==13.9.4
rlp==4.0.1
rootpath==0.1.1
rpds-py==0.22.3
rsa==4.9
s3transfer==0.11.1
safetensors==0.5.2
schedule==1.2.2
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.15.0
sdnotify==0.3.2
semantic-version==2.10.0
sentencepiece==0.2.0
seqeval==1.2.2
setproctitle==1.3.4
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
smart_open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.5.1
starlette==0.41.3
statsforecast==1.7.8
statsmodels==0.14.4
sympy==1.13.3
TA-Lib==0.4.32
tabulate==0.9.0
technical==1.5.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard_data_server==0.7.0
tensorboardX==2.6.2.2
termcolor==2.5.0
text-unidecode==1.3
textual==1.0.0
thinc==8.3.2
threadpoolctl==3.5.0
tifffile==2024.12.12
timm==1.0.3
tokenizers==0.21.0
toolz==0.12.1
torch==2.3.1
torchmetrics==1.2.1
torchvision==0.15.2a0
tox==4.23.2
tqdm==4.67.1
transformers==4.48.0
triad==0.9.8
trie==3.0.1
typer==0.15.1
typer-slim==0.15.1
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
uc-micro-py==1.0.3
unicodedata2==16.0.0
urllib3==2.2.3
utilsforecast==0.2.3
uvicorn==0.34.0
virtualenv==20.28.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.4.1
web3==7.6.0
websockets==13.1
Werkzeug==3.1.3
wheel==0.44.0
window_ops==0.0.15
wrapt==1.17.2
xgboost==2.1.3
xxhash==3.5.0
yarl==1.9.4
zipp==3.21.0
zstandard==0.23.0
